feat: update data to support biginter

+++ src/api/models/data.py
@@ -1,6 +1,6 @@
 from decimal import Decimal
 
-from sqlalchemy import DECIMAL, Integer
+from sqlalchemy import DECIMAL, BigInteger
 from sqlalchemy.orm import Mapped, mapped_column
 
 from src.infrastructure.database import Base
@@ -10,7 +10,7 @@ class Data(Base):
     __tablename__ = "data"
 
     id: Mapped[int] = mapped_column(primary_key=True, unique=True)
-    timestamp: Mapped[int] = mapped_column(Integer)
+    timestamp: Mapped[int] = mapped_column(BigInteger)
     wind_speed: Mapped[Decimal] = mapped_column(DECIMAL(10, 2))
     power: Mapped[Decimal] = mapped_column(DECIMAL(10, 2))
     ambient_temperature: Mapped[Decimal] = mapped_column(DECIMAL(10, 2))
